feat: add update and delete tasks

+++ task_cli.py
@@ -25,11 +25,10 @@ elif sys.argv[1] == "add": # Add a new task / Añadir una nueva tarea
         error_message("Add command requires at least one argument.") # El comando Add requiere al menos un argumento.
     elif len(sys.argv) < 3:
         error_message("The task description is mandatory.") # La descripción de la tarea es obligatoria.
-    elif len(sys.argv) > 3:
-        error_message("Add command only accepts a description as its sole argument.") # El comando Add solo acepta una descripción como su único argumento.
     else:
+        complete_description = " ".join(sys.argv[2:])
         if not os.path.exists("to-do-list.json"):
-            task_properties = [add_task("1", sys.argv[2])]
+            task_properties = [add_task("1", complete_description)]
             y = json.dumps(task_properties, indent=4)
             with open("to-do-list.json", "w") as f:
                 f.write(y)
@@ -37,26 +36,74 @@ elif sys.argv[1] == "add": # Add a new task / Añadir una nueva tarea
         else:
             with open("to-do-list.json", "r") as f:
                 data = json.load(f)
-                task_list = []
-                for task in data:
-                    task_id = int(task["id"])
-                    task_list.append(task_id)
-                new_id = str(max(task_list) + 1)
-                task_properties = add_task(new_id, sys.argv[2])
-                data.append(task_properties)
-                y = json.dumps(data, indent=4)
-            with open("to-do-list.json", "w") as f:
-                f.write(y)
-            print(f"Task added successfully (ID: {new_id})")
+                if data == []: # If the file is empty but it exists/ Si el archivo está vacío pero existe
+                    new_id = "1"
+                    task_properties = add_task(new_id, complete_description)
+                    data.append(task_properties)
+                    y = json.dumps(data, indent=4)
+                    with open("to-do-list.json", "w") as f:
+                        f.write(y)
+                    print(f"Task added successfully (ID: {new_id})")
+                    sys.exit(0)
+                else:
+                    task_list = []
+                    for task in data:
+                        task_id = int(task["id"])
+                        task_list.append(task_id)
+                    new_id = str(max(task_list) + 1)
+                    task_properties = add_task(new_id, complete_description)
+                    data.append(task_properties)
+                    y = json.dumps(data, indent=4)
+                with open("to-do-list.json", "w") as f:
+                    f.write(y)
+                print(f"Task added successfully (ID: {new_id})")
 
-elif sys.argv[1] == "update":
-    print("Se ha eliminado la tarea llamada." + sys.argv[2])
-elif sys.argv[1] == "delete":
-    print("Mostrando todas las tareas.")
+elif sys.argv[1] == "update": # Update a new task / Actualizar una tarea
+    if len(sys.argv) < 4:
+        error_message("Update command requires a task ID and a new description.") # El comando Update requiere un ID de tarea y una nueva descripción.
+    else:
+        with open("to-do-list.json", "r") as f:
+            data = list(json.load(f))
+            task_found = False
+            for task in data:
+                if task["id"] == sys.argv[2]:
+                    new_description = " ".join(sys.argv[3:])
+                    task["description"] = new_description
+                    task["UpdatedAt"] = datetime.datetime.now().isoformat() + "Z"
+                    print(f"Task updated successfully (ID: {sys.argv[2]})")
+                    task_found = True
+                    y = json.dumps(data, indent=4)
+                    with open("to-do-list.json", "w") as f:
+                        f.write(y)
+                    break
+            if not task_found:
+                error_message(f"Task with ID {sys.argv[2]} not found.")    
+
+elif sys.argv[1] == "delete": # Delete a task / Eliminar una tarea
+    if len(sys.argv) < 3:
+        error_message("Delete command requires a task ID.") # El comando Delete requiere un ID de tarea.
+    else:
+        with open("to-do-list.json", "r") as f:
+            data = list(json.load(f))
+            task_found = False
+            for task in data:
+                if task["id"] == sys.argv[2]:
+                    data.remove(task)
+                    print(f"Task deleted successfully (ID: {sys.argv[2]})")
+                    task_found = True
+                    y = json.dumps(data, indent=4)
+                    with open("to-do-list.json", "w") as f:
+                        f.write(y)
+                    break
+            if not task_found:
+                error_message(f"Task with ID {sys.argv[2]} not found.")
+
 elif sys.argv[1] == "mark-in-progress":
     print("Marcando la tarea como en progreso." + sys.argv[2])
+
 elif sys.argv[1] == "mark-done":
     print("Marcando la tarea como completada." + sys.argv[2])
+
 elif sys.argv[1] == "list":
     if len(sys.argv) == 3:
         if sys.argv[2] == "done":
